Solution

+++ app/create_file.py
@@ -1,62 +1,47 @@
-import sys
 import os
+import argparse
 from datetime import datetime
+
+
+def main(dir_path: list[str] = None, file_name: str = None) -> None:
+    if dir_path:
+        path = os.path.join(*dir_path)
+        os.makedirs(path, exist_ok=True)
+    if file_name:
+        if dir_path:
+            file_path = os.path.join(*dir_path, file_name)
+            data_input(file_path)
+        else:
+            data_input(file_name)
 
 
 def data_input(file_path: str) -> None:
     count = 0
+    if os.path.exists(file_path):
+        with open(file_path) as file:
+            count = len(file.readlines()) - 1
+
     utc_time = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-    with open(file_path, "w") as file_time:
-        file_time.write(utc_time + "\n")
-        file_time.close()
     while True:
         count += 1
         text = input("Enter content line: ")
         if text.lower() == "stop":
             break
-
-        with open(file_path, "a") as file:
-            file.write(f"{count} {text} \n")
-
-
-def only_d() -> None:
-    if "-d" in sys.argv and "-f" not in sys.argv:
-        d_index = sys.argv.index("-d")
-        list_d = []
-        list_d.extend(sys.argv[d_index + 1:])
-        dir_path = os.path.join(list_d)
-        if dir_path:
-            os.makedirs(dir_path, exist_ok=True)
+        if not os.path.exists(file_path):
+            with open(file_path, "w") as file_time:
+                file_time.write(utc_time + "\n")
+                file_time.write(f"{count} {text} \n")
+        else:
+            with open(file_path, "a") as file:
+                file.write(f"{count} {text} \n")
 
 
-def only_f() -> None:
-    if "-f" in sys.argv and "-d" not in sys.argv:
-        f_index = sys.argv.index("-f")
-        utc_time = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
+parser = argparse.ArgumentParser(description="Create file with input")
 
-        with open(sys.argv[f_index + 1], "w") as file_time:
-            file_time.write(utc_time + "\n")
-            file_time.close()
-        data_input(sys.argv[f_index + 1])
+parser.add_argument("-d", nargs="+", dest="dir_path")
+parser.add_argument("-f", dest="file_name")
 
+args = parser.parse_args()
 
-def with_d_and_f() -> None:
-    if "-d" in sys.argv and "-f" in sys.argv:
-        d_index = sys.argv.index("-d")
-        f_index = sys.argv.index("-f")
-
-        list_d = []
-
-        list_d.extend(sys.argv[d_index + 1:f_index])
-        dir_path = os.path.join(*list_d)
-        if dir_path:
-            os.makedirs(dir_path, exist_ok=True)
-
-        file_path = os.path.join(dir_path, sys.argv[f_index + 1])
-
-        data_input(file_path)
-
-
-with_d_and_f()
-only_d()
-only_f()
+if __name__ == "__main__":
+    main(args.dir_path, args.file_name)
